Update to typescript 3.7.3, node 12.4
VSCode 1.41 now officially supports it.
And, we need the newer node types files (due to an bug in the 10.11 ones), so
we might as well move to node 12.4, which vscode 1.40.0+ has.

The new node types also updated some fields to optional, so some code needed to
be updated to handle null/undefined.

And, added a helper script to build the .vsix file locally, since I was getting
tired of typing it all out.