Fix intermittent TemplateEditorManager test fails on linux
On the linux test runs, it seems something is opening the temp test template
files, triggering editing support to be configured, which the test is expecting
to not be initially setup.
However, we don't really care about when editing is configured, just that it is
definitely configured when the file is opened, so I just removed the initial
checks.