Allows literal nulls in fields that support it
Once you allow null in a json-schema on otherwise string types, VSCode stops
doing the default text-based hippy completion for the fields; it also doesn't
offer the type-specific empty value (e.g. "", [], {}) when null is allowed for
other types.
So, this adds in those empty values, via either "examples" or "defaultSnippets"
in the schema.