Warn on duplicate rel paths configured in a template
With the way we do the schema assignments, the relpath file will get
that last schema configured, which might not give the best warnings, but
the warning in template-info.json should be enough to point out that
something's wrong.
Also, no relpath should be 'template-info.json' nor should that appear
in the code completions, since that will never be valid.