Escape codicons in server response text
In the quickpick api, you can use $(...) in the ui labels to add icons, so we
want to escape any text from the server that might include $(...) so it shows as
the original text not the icon.
Also, added unit test for analyticsdx-vscode-core for this, which meant adding
the unit test pattern to the project too.