Use vscode-json-languageserver module
Instead of using the one from the vscode.json-language-features extension in the
vscode install. In insiders 1.47, that server code in the extension changed
enough that our extension wasn't able to start it up anymore.
This gives us a known version of the json language server to work with so we're
independent from the vscode (and lsp) version.