Fix json language server in vscode 1.47+ (#98)
* Use vscode-json-languageserver module

Instead of using the one from the vscode.json-language-features extension in the
vscode install. In insiders 1.47, that server code in the extension changed
enough that our extension wasn't able to start it up anymore.
This gives us a known version of the json language server to work with so we're
independent from the vscode (and lsp) version.

* Fix test:vscode-insiders-integration

So it works on all platforms, and hook it into our github actions builds so we know earlier if the extension are going to fail in the next version of vscode.