Show loading and no items message in pickers (#109)
Now, for the commands that show a list of apps or templates, it will show the empty picker right away with the loading bar and a loading message. Once the items load, you can pick one, or, if there are none available, it will show a message in the picker (instead of hiding the picker and showing a notification popup in the bottom corner, which was kind of jarring).