Add template json schema validation to lint package
This moves the template json schemas from the vscode extension into the lint
package, and updates the extension to reference them out of the package.
It also moves schema-specific unit tests to the package, but leaves tests
that verify the vscode portions of and usages of the schemas in the extension.